Align RPC class names with client-java
